test_match v0.1.0 TestMatch.Lib

Module for reading and writing a library of libgraph graphs.

The read_graph/1 and write_graph/2 functions allow for reading and writing RDF graphs to the project data repository. (Default file names are provided with the read_graph/0 and write_graph/1 forms.) The list_graphs/0 function lists graph file names.

Some simple accessor functions are also available:

graphs_dir/0, graph_file/0, temp_graph_file/0

Functions

create_ets_tables()

Create ETS tables

from_cypher(query \\ "match (n)-[r]->(o) return n,r,o")

Executes cypher query and returns a graph.

from_cypher_with_properties(query \\ "match (n)-[r]->(o) return n,r,o")

Executes cypher query and returns a graph with properties in ETS tables.

from_sparql(query \\ "select * where {?s ?p ?o}")

graph_file()

Returns a default Lib graph file.

Examples

iex> TestMatch.Lib.graph_file()
"default.ttl"

graphs_dir()

Returns the Lib graphs directory.

list_graph_images()

Lists Lib graphs in the Lib graph images library.

Examples

iex> list_graph_images()
["tony.png", "foo.png", "foo1.png"]

list_graphs()

Lists Lib graphs in the Lib graphs library.

Examples

iex> list_graphs()
["tony.dot", "foo.dot", "foo1.dot"]

read_graph(graph_file \\ graph_file())

Reads a Lib graph from the Lib graphs library.

Examples

iex> read_graph()
%TestMatch.Graph{
  data: "<http://dbpedia.org/resource/Hello_World>\n..."
  file: "default.ttl",
  path: ".../test_graph/priv/rdf/graphs/default.ttl",
  type: :rdf,
  uri: "file:///.../test_graph/priv/rdf/graphs/default.ttl"
}

iex> read_graph("books.ttl")
%TestMatch.Graph{
  data: "@prefix bibo: <http://purl.org/ontology/bibo/> \n..."
  file: "books.ttl",
  path: ".../test_graph/priv/rdf/graphs/books.ttl",
  type: :rdf,
  uri: "file:///.../test_graph/priv/rdf/graphs/books.ttl"
}

temp_graph_file()

Returns a temp Lib graph file for writing.

Examples

iex> TestMatch.Lib.temp_graph_file()
"temp.ttl"

to_png(graph, binary \\ :dot)

Writes a graph to a .png file in the Lib graph images library.

:dot − filter for drawing directed graphs :neato − filter for drawing undirected graphs :twopi − filter for radial layouts of graphs :circo − filter for circular layout of graphs :fdp − filter for drawing undirected graphs :sfdp − filter for drawing large undirected graphs :patchwork − filter for squarified tree maps :osage − filter for array-based layouts

write_graph(graph_data, graph_file \\ temp_graph_file())

Writes a Lib graph to a file in the Lib graphs library.

Examples

iex> data |> write_graph("my.dot")
%TestMatch.Graph{
  data: "@prefix bibo: <http://purl.org/ontology/bibo/> \n..."
  file: "my.dot",
  path: "/test_graph/priv/lib/graphs/my.dot",
  type: :lib,
  uri: "file:///.../test_graph/priv/lib/graphs/my.dot"
}

write_lib_graph(lib_graph, graph_file \\ temp_graph_file())

Writes a Lib graph to a file in the Lib graphs library.